+++ tui/src/cli.rs
@@ -27,10 +27,8 @@ use std::path::PathBuf;
 #[derive(Parser, Debug)]
 // mostly read from `Cargo.toml`
 #[clap(name = "Termusic", author, version = env!("TERMUSIC_VERSION"), about, long_about=None)]
-// #[clap(next_line_help = true)]
-// #[clap(propagate_version = true)]
 pub struct Args {
-    /// Commands for podcast
+    /// Subcommands, overwriting default action of starting the TUI
     #[command(subcommand)]
     pub action: Option<Action>,
     /// With no `MUSIC_DIRECTORY`, use config in `~/.config/termusic/config.toml`,
@@ -82,14 +80,15 @@ impl std::fmt::Display for Backend {
     }
 }
 
+/// Subcommands for the binary
 #[derive(Subcommand, Debug)]
 pub enum Action {
-    /// Import feeds from opml file.
+    /// Import Podcast feeds from a opml file.
     Export {
         #[arg(value_name = "FILE")]
         file: PathBuf,
     },
-    /// Export feeds to opml file.
+    /// Export Podcast feeds to a opml file.
     Import {
         #[arg(value_name = "FILE")]
         file: PathBuf,